refactor: remove tool base command
